Update: Move metadata service to client
